refactor(registry): split jobs matchingHelpers under 200-line policy
